PlotSpecActivity_fixes
Line 775 does not work properly when the user selects specs (e.g. spec_list = c(1,4)). If c is place where the retdata are saved, the length(levels(c$spec)) gives result zero, when spec_list  is comprised of numbers. As result the if condition returns false, and the rest of commands will not run. To overcome this problem, I introduced the lines 740-741, so that the systems takes the numbers of spec_list and returns their names as strings. Then the function works as previously.  The functionality of line 775 is transferred in the graph g1_5, as the latter is the same as q2 but contains "+ ggplot2::facet_wrap(~spec, scales="free_y")" to plot the variance for each spec. The q2 remains intact because (when 775 did not work) it used to plot the overall variance, as @euba proposed in the issue https://github.com/euba/BacArena/issues/60 . The information of g2 may be of interest to researchers in the future.