Substances in plotReaActivity?
In my opinion, there is a mishmash of substances and reactions in plotReaActivity. I assume that the author wanted to propose the number of reactions in line 839, not the number of substances. In "df" there should be no "sub" information. In 815 a similar mishmash may have taken place, but at least this is just a description